Fix Visual Studio compiler warning (strcat and sscanf)
This is a follow-up to commits 492ee05a6b and 4b9e9b1d28

Visual Studio complains:
  .\file_xfer.cpp(90) : warning C4996: 'strcat': This function or variable
  may be unsafe. Consider using strcat_s instead.
  To disable deprecation, use _CRT_SECURE_NO_WARNINGS.
  See online help for details.

And a similar complain for sscanf.

Replace them with the secure function when compiling with Visual Studio.
Note that the size provided is sizeof(d), which means all calls to
sscanf and strcat must be done with the first param (destination) being
an array and not a pointer.